% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msqrob-utils.R
\name{checkReference}
\alias{checkReference}
\title{Check for presence of reference levels}
\usage{
checkReference(y, data, paramNames, factorLevels)
}
\arguments{
\item{y}{A numeric vector with the feature values for each sample.}

\item{data}{A \code{DataFrame} with information on the design. It has
the same number of rows as the number of columns (samples) of
\code{y}.}

\item{paramNames}{A \code{character} vector with the names of the parameters in the model.}

\item{factorLevels}{A list with the levels of the factors in the model. The names of the list
should correspond to the factor variables present in the model.}
}
\value{
A logical vector indicating for each parameter (that comes from a factor variable) in the model
if his reference level has values for the current feature.
}
\description{
Check if the reference levels of the factors in the model are present for a specific feature.
}
