% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msqrob-utils.R
\name{getFormulaFactors}
\alias{getFormulaFactors}
\title{Get the levels of the factors in the model}
\usage{
getFormulaFactors(formula, dataFrame)
}
\arguments{
\item{formula}{A \code{formula} object that specifies the model.}

\item{data}{A \code{DataFrame} with information on the design. Contains the variables that are used in the model.}
}
\value{
A list with the levels of the factors in the model. The names of the list
correspond to the factor variables present in the model.
}
\description{
Get the levels of the factors in the model associated with the factors
names.
}
