% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msqrob-utils.R
\name{getReferenceLevels}
\alias{getReferenceLevels}
\title{Get reference levels factors}
\usage{
getReferenceLevels(dataFrame, formula)
}
\arguments{
\item{dataFrame}{A \code{data.frame} that contains factors}

\item{formula}{A formula object associated with the data frame (the variables in the formula should be present in the data frame)}
}
\value{
A named list with the reference levels of the factors (named by the factor name)
}
\description{
Get the reference levels of the factors of a data frame that are present in a formula
}
