% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msqrob-framework.R
\docType{class}
\name{StatModel-class}
\alias{StatModel-class}
\alias{show,StatModel-method}
\title{The StatModel class for msqrob}
\usage{
\S4method{show}{StatModel}(object)
}
\description{
classes for msqrob
}
\section{Slots}{

\describe{
\item{\code{type}}{The type of the used model}

\item{\code{params}}{A list containing information of the used model}

\item{\code{varPosterior}}{A vector of posterior variance}

\item{\code{dfPosterior}}{A vecotr of posterior degrees of freedom}
}}

\examples{
myModel <- new("StatModel", type = "rlm", params = list(x = 3, y = 7, b = 4), varPosterior = c(0.1, 0.2, 0.3), dfPosterior = c(6, 7, 8))
myModel@params
mod1 <- new("StatModel", type = "rlm")
mod2 <- new("StatModel", type = "lm")
df <- DataFrame(x = 1:2)
df$mods <- c(mod1, mod2)
df$mods
}
\author{
Oliver M. Crook, Lieven Clement
}
