% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msqrob.R, R/msqrobQB.R
\name{msqrobGlm}
\alias{msqrobGlm}
\alias{msqrobQB,Features-method}
\title{Function to fit msqrob models to peptide counts using glm}
\usage{
msqrobGlm(y, npep, formula, data, priorCount = 0.1, binomialBound = TRUE)

\S4method{msqrobQB}{Features}(
  object,
  i,
  formula,
  modelColumnName = "msqrobQbModels",
  overwrite = FALSE,
  priorCount = 0.1,
  binomialBound = TRUE
)
}
\arguments{
\item{y}{A \code{matrix} with the peptide counts. The
features are along the rows and samples along the columns.}

\item{formula}{Model formula. The model is built based on the
covariates in the data object.}

\item{data}{A \code{DataFrame} with information on the design. It has
the same number of rows as the number of columns (samples) of
\code{y}.}

\item{priorCount}{A 'numeric(1)', which is a prior count to be added to the observations to shrink
the estimated log-fold-changes towards zero.}

\item{i}{\code{character} or \code{integer} to specify the element of the \code{Features} that
contains the log expression intensities that will be modelled.}

\item{nPep}{A vector with number of peptides per protein. It has as length
the number of rows of y. The counts are equal or larger than the largest
peptide count in y.}

\item{binomialBound:}{logical, if ‘TRUE’ then the quasibinomial variance estimator will
be never smaller than 1 (no underdispersion).}
}
\value{
A list of objects of the \code{StatModel} class.
}
\description{
Low-level function for parameter estimation with msqrob
by modeling peptide counts using quasibinomial glm
}
\examples{


}
\author{
Lieven Clement
}
