% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msqrob-utils.R
\name{referenceContrast}
\alias{referenceContrast}
\title{Check if the reference levels associated with the contrast are present}
\usage{
referenceContrast(referencePresent, L, acceptDifferentReference)
}
\arguments{
\item{referencePresent}{A logical vector indicating for each parameter in the model
if his reference level has values for the current feature.}

\item{L}{A numeric contrast matrix with rownames that equal the model parameters
that are involved in the contrasts}

\item{acceptDifferentReference}{\code{boolean(1)} to indicate if the contrasts that involve
parameters with modified reference levels are returned.}
}
\value{
A boolean
}
\description{
Check if the reference levels associated with the contrast are present
}
