% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessors.R
\name{getModel,StatModel-method}
\alias{getModel,StatModel-method}
\alias{statModelAccessors}
\alias{getModel}
\alias{getFitMethod}
\alias{getCoef}
\alias{getDF}
\alias{getDfPosterior}
\alias{getVarPosterior}
\alias{getSigmaPosterior}
\alias{getVar}
\alias{getSigma}
\alias{getVcovUnscaled}
\title{Accessor functions for StatModel class}
\usage{
\S4method{getModel}{StatModel}(object)
}
\arguments{
\item{object}{\code{StatModel} object}
}
\description{
Accessor functions for StatModel class
\describe{
\item{getModel(object)}{to get model}
\item{getFitMethod(object)}{to get the parameter estimation method}
\item{getCoef(object)}{to get the parameter estimates of the mean model}
\item{getDF(object)}{to get the residual degrees of freedom of the model}
\item{getVar(object)}{to get the residual variance of the model}
\item{getSigma(object)}{to get the residual standard deviation of the model}
\item{getDfPosterior(object)}{to get the degrees of freedom of
the empirical Bayes variance estimator}
\item{getVarPosterior(object)}{to get the empirical Bayes variance}
\item{getSigmaPosterior(object)}{to get the empirical Bayes standard deviation}
\item{getVcovUnscaled(object)}{to get the unscaled variance covariance matrix
of the model parameters}
}
}
