% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msqrob-utils.R
\name{checkReference}
\alias{checkReference}
\title{Check for presence of reference levels}
\usage{
checkReference(y, data, referenceLevels)
}
\arguments{
\item{y}{A numeric vector with the feature values for each sample.}

\item{data}{A \code{DataFrame} with information on the design. It has
the same number of rows as the number of columns (samples) of
\code{y}.}

\item{formula}{Model formula. The model is built based on the
covariates in the data object.}

\item{paramNames}{character vector specifying the model parameters that are present in the global model matrix}
}
\value{
A logical vector indicating for each parameter in the model if his reference level has values for the current feature.
}
\description{
Check if the reference levels of the factors in the model are present for a specific feature.
}
