% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hypothesisTest-methods.R
\name{hypothesisTest,SummarizedExperiment-method}
\alias{hypothesisTest,SummarizedExperiment-method}
\alias{hypothesisTest}
\alias{hypothesisTest,Features-method}
\alias{msqrobAggregate,Features-method}
\title{Parameter estimates, standard errors and statistical inference on differential
expression}
\usage{
\S4method{hypothesisTest}{SummarizedExperiment}(
  object,
  contrast,
  adjust.method = "BH",
  modelColumn = "msqrobModels",
  resultsColumnName = "msqrobTestResults",
  overwrite = FALSE
)

\S4method{hypothesisTest}{Features}(
  object,
  i,
  contrast,
  adjust.method = "BH",
  modelColumn = "msqrobModels",
  resultsColumnName = "msqrobTestResults",
  overwrite = FALSE
)
}
\arguments{
\item{object}{\code{SummarizedExperiment} or \code{Features} instance}

\item{contrast}{\code{numeric} matrix specifying one or more contrasts of
the linear model coefficients to be tested equal to zero.
The rownames of the matrix should be equal to the names of
parameters of the model.}

\item{adjust.method}{\code{character} specifying the method to adjust
the p-values for multiple testing.
Options, in increasing conservatism, include ‘"none"’,
‘"BH"’, ‘"BY"’ and ‘"holm"’.  See ‘p.adjust’ for the complete
list of options. Default is "BH" the Benjamini-Hochberg method
to controle the False Discovery Rate (FDR).}

\item{modelColumn}{\code{character} to indicate the variable name that is used
to store the msqrob models in the rowData of the SummarizedExperiment
instance or of the assay of the Features instance. Default is "msqrobModels".}

\item{resultsColumnName}{\code{character} to indicate the variable name that will be used
to store test results in the rowData of the SummarizedExperiment instance
or of the assay of the Features instance. Default is "msqrobTestResults".}

\item{i}{\code{character} or \code{integer} to specify the element of the \code{Features} that
contains the log expression intensities that will be modelled.}
}
\value{
A SummarizedExperiment or a \code{Features} instance augmented with the test
results.
}
\description{
Summary table of the estimates for differential expression of features
}
\examples{

# Load example data
# The data are a Feature object with containing
# a SummarizedExperiment named "peptide" with MaxQuant peptide intensities
# The data are a subset of spike-in the human-ecoli study
# The variable condition in the colData of the Feature object
# contains information on the spike in condition a-e (from low to high)
data(pe)

# Aggregate peptide intensities in protein expression values
pe<-aggregateFeatures(pe,i="peptide",fcol="Proteins",name="protein")

# Fit msqrob model
pe <- msqrob(pe,i="protein",formula=~condition)

#Define contrast
coef1<-getCoef(rowData(pe[["protein"]])$rlm[[1]])
L <- matrix(0,ncol=1,nrow=length(coef1))
rownames(L) <- names(coef1)
L[2,1]<-1

#example SummarizedExperiment instance
se <- pe[["protein"]]
se <- hypothesisTest(se,L)
head(rowData(se)$msqrobTestResults,10)

#example Features instance
pe <- hypothesisTest(pe,i="protein",L)
head(rowData(pe[["protein"]])$msqrobTestResults,10)
#Volcano plot
plot(-log10(pval)~logFC,rowData(pe[["protein"]])$msqrobTestResults,col=(adjPval<0.05)+1)
}
\author{
Lieven Clement
}
