% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msqrob.R
\name{msqrobLm}
\alias{msqrobLm}
\title{Function to fit msqrob models}
\usage{
msqrobLm(y, formula, data, robust = TRUE, maxitRob = 5)
}
\arguments{
\item{y}{is a matrix with the quantified MS intensities, the features are in the rows and samples in the columns.}

\item{formula}{Model formula. The model is built based on the covariates in the data object.}

\item{data}{is a data frame of class DataFrame with information on the design. It has the same number of rows as the number of columns (samples) of y.}

\item{robust}{boolean to indicate if robust regression is performed to account for outliers. If 'FALSE' an OLS fit is performed.}

\item{maxitRob}{Maximum iterations in the IRWLS algorithm used in the M-estimation step of the robust regression.}
}
\value{
A list of objects of the StatModel class
}
\description{
Parameter estimation of msqrob models.
}
\examples{
#TODO
}
\author{
Lieven Clement, Oliver M. Crook
}
