% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msqrob.R
\name{msqrob}
\alias{msqrob}
\title{Function to fit msqrob models}
\usage{
msqrob(pe, assay, formula, robust = TRUE, maxitRob = 5)
}
\arguments{
\item{pe}{Proteomics Experiment Object, is an object off class Feature set that contains the assay with the quantified MS intensities.}

\item{assay}{integer or string referring to the assay of the Feature object containing the intensities for modelling, e.g. peptide or protein.}

\item{formula}{Model formula. The model is built based on the covariates in the colData of the pe object.}

\item{robust}{bolean to indicate if robust regression is performed to account for outliers. If 'FALSE' an OLS fit is performed.}

\item{maxitRob}{Maximum iterations in the IRWLS algorithm used in the M-estimation step of the robust regression.}
}
\value{
A list of objects of the StatModel class
}
\description{
Parameter estimation of msqrob models.
}
\author{
Lieven Clement, Oliver M. Crook
}
