% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collect-plot-results.R
\name{msqrobCollect}
\alias{msqrobCollect}
\title{Function to collect the inference tables generated by the msqrob2
statistical inference workflow.}
\usage{
msqrobCollect(object, contrast, resultsColumnNamePrefix = "", combine = TRUE)
}
\arguments{
\item{object}{A \code{SummarizedExperiment} object containing model
inference performed by \code{hypothesisTest()}.}

\item{contrast}{A numeric \code{matrix} providing the contrast matrix
used during model estimation or a character \code{vector} providing
the names of the contrast used during model estimation.}

\item{resultsColumnNamePrefix}{A \code{character(1)} providing the
prefix used when running \code{hypothesisTest()}.}

\item{combine}{A \code{logical(1)} indicating whether the result tables
should be combined in a single table (default) or return as a
list of tables. When combined, two new variables are created:
1. \code{contrast}tells from which contrast (column of the \code{contrast}
matrix) the results where obtained; 2. \code{feature} provides the
name of the modelled feature (this information is taken from
the rownames of the tables, but this are made unique upon
combining).}
}
\value{
Result tables for the contrasts combined in a single inference table or list
}
\description{
Function to collect the inference tables generated by the msqrob2
statistical inference workflow.
}
